Refactor core structures into modules
