+++ src/lib.rs
@@ -1,18 +1,22 @@
-use bs58;
 pub use coin_proto::{Transaction, TransactionInput, TransactionOutput};
-use contract_runtime;
-use ripemd::Ripemd160;
-use secp256k1;
-use serde_json;
 use sha2::{Digest, Sha256};
-use std::collections::HashMap;
-use std::fs::File;
-use std::io::Write;
-use std::path::Path;
-use std::time::{SystemTime, UNIX_EPOCH};
 
+pub mod block;
+pub mod blockchain;
 pub mod blockfile;
+pub mod transaction;
 pub mod utils;
+
+pub use block::{Block, BlockHeader};
+pub use blockchain::{
+    BLOCK_SUBSIDY, Blockchain, COIN, DIFFICULTY_WINDOW, HALVING_INTERVAL, MAX_SUPPLY,
+    TARGET_BLOCK_MS,
+};
+pub use transaction::{
+    Error, Result, TransactionExt, address_from_secret, coinbase_transaction, decrypt_message,
+    encrypt_message, new_multi_transaction_with_fee, new_transaction, new_transaction_with_fee,
+    new_transaction_with_message, valid_address,
+};
 pub mod utxofile;
 pub use utils::meets_difficulty;
 
@@ -46,739 +50,6 @@ pub fn compute_merkle_root(txs: &[Transaction]) -> String {
     merkle_root_from_hashes(&hashes)
 }
 
-/// Number of blocks used for difficulty adjustment
-pub const DIFFICULTY_WINDOW: usize = 3;
-/// Target time between blocks in milliseconds
-pub const TARGET_BLOCK_MS: u64 = 400;
-/// Smallest unit of the coin (1 coin = 100 million units)
-pub const COIN: u64 = 100_000_000;
-/// Initial reward paid to miners for producing a block
-pub const BLOCK_SUBSIDY: u64 = 50 * COIN;
-/// Number of blocks between reward halvings. With an initial subsidy of 50 this
-/// caps total issuance at 20 million coins.
-pub const HALVING_INTERVAL: u64 = 200_000;
-/// Maximum number of units that will ever exist
-pub const MAX_SUPPLY: u64 = 20_000_000 * COIN;
-
-#[derive(Debug, PartialEq)]
-pub enum Error {
-    InvalidAddress,
-}
-
-pub type Result<T> = std::result::Result<T, Error>;
-
-pub fn new_transaction_with_fee(
-    sender: impl Into<String>,
-    recipient: impl Into<String>,
-    amount: u64,
-    fee: u64,
-) -> Result<Transaction> {
-    let sender = sender.into();
-    let recipient = recipient.into();
-    if !valid_address(&sender) || !valid_address(&recipient) {
-        return Err(Error::InvalidAddress);
-    }
-    Ok(Transaction {
-        sender,
-        recipient,
-        amount,
-        fee,
-        signature: Vec::new(),
-        encrypted_message: Vec::new(),
-        inputs: Vec::new(),
-        outputs: Vec::new(),
-    })
-}
-
-pub fn new_transaction(
-    sender: impl Into<String>,
-    recipient: impl Into<String>,
-    amount: u64,
-) -> Result<Transaction> {
-    new_transaction_with_fee(sender, recipient, amount, 0)
-}
-
-pub fn new_multi_transaction_with_fee(
-    sender: impl Into<String>,
-    outputs: Vec<(String, u64)>,
-    fee: u64,
-) -> Result<Transaction> {
-    let sender = sender.into();
-    if !valid_address(&sender) {
-        return Err(Error::InvalidAddress);
-    }
-    let recipient = outputs.first().map(|o| o.0.clone()).unwrap_or_default();
-    let amount = outputs.first().map(|o| o.1).unwrap_or(0);
-    let extra: Vec<_> = outputs
-        .into_iter()
-        .skip(1)
-        .map(|(addr, amt)| TransactionOutput {
-            address: addr,
-            amount: amt,
-        })
-        .collect();
-    if !valid_address(&recipient) || extra.iter().any(|o| !valid_address(&o.address)) {
-        return Err(Error::InvalidAddress);
-    }
-    Ok(Transaction {
-        sender,
-        recipient,
-        amount,
-        fee,
-        signature: Vec::new(),
-        encrypted_message: Vec::new(),
-        inputs: Vec::new(),
-        outputs: extra,
-    })
-}
-
-/// Create a coinbase transaction paying `amount` to `miner`
-pub fn coinbase_transaction(miner: impl Into<String>, amount: u64) -> Result<Transaction> {
-    let miner = miner.into();
-    if !valid_address(&miner) {
-        return Err(Error::InvalidAddress);
-    }
-    Ok(Transaction {
-        sender: String::new(),
-        recipient: miner,
-        amount,
-        fee: 0,
-        signature: Vec::new(),
-        encrypted_message: Vec::new(),
-        inputs: Vec::new(),
-        outputs: Vec::new(),
-    })
-}
-
-pub fn encrypt_message(msg: &str, sk: &secp256k1::SecretKey, pk: &secp256k1::PublicKey) -> Vec<u8> {
-    let secret = secp256k1::ecdh::SharedSecret::new(pk, sk);
-    let key = Sha256::digest(secret.as_ref());
-    msg.as_bytes()
-        .iter()
-        .enumerate()
-        .map(|(i, b)| b ^ key[i % key.len()])
-        .collect()
-}
-
-pub fn decrypt_message(
-    data: &[u8],
-    sk: &secp256k1::SecretKey,
-    pk: &secp256k1::PublicKey,
-) -> Option<String> {
-    let secret = secp256k1::ecdh::SharedSecret::new(pk, sk);
-    let key = Sha256::digest(secret.as_ref());
-    let bytes: Vec<u8> = data
-        .iter()
-        .enumerate()
-        .map(|(i, b)| b ^ key[i % key.len()])
-        .collect();
-    String::from_utf8(bytes).ok()
-}
-
-pub fn new_transaction_with_message(
-    sender: impl Into<String>,
-    recipient: impl Into<String>,
-    amount: u64,
-    fee: u64,
-    message: &str,
-    sender_sk: &secp256k1::SecretKey,
-    recipient_pk: &secp256k1::PublicKey,
-) -> Result<Transaction> {
-    let sender = sender.into();
-    let recipient = recipient.into();
-    if !valid_address(&sender) || !valid_address(&recipient) {
-        return Err(Error::InvalidAddress);
-    }
-    Ok(Transaction {
-        sender,
-        recipient,
-        amount,
-        fee,
-        signature: Vec::new(),
-        encrypted_message: encrypt_message(message, sender_sk, recipient_pk),
-        inputs: Vec::new(),
-        outputs: Vec::new(),
-    })
-}
-
-/// Validate that an address is Base58Check encoded and 34 characters long.
-pub fn valid_address(addr: &str) -> bool {
-    if addr.len() != 34 {
-        return false;
-    }
-    let bytes = match bs58::decode(addr).into_vec() {
-        Ok(b) => b,
-        Err(_) => return false,
-    };
-    if bytes.len() != 25 || bytes[0] != 0x00 {
-        return false;
-    }
-    let (payload, checksum) = bytes.split_at(21);
-    let check = Sha256::digest(Sha256::digest(payload));
-    checksum == &check[..4]
-}
-
-/// Derive an address from a secret key.
-pub fn address_from_secret(sk: &secp256k1::SecretKey) -> String {
-    let secp = secp256k1::Secp256k1::new();
-    let pk = secp256k1::PublicKey::from_secret_key(&secp, sk);
-    let pk_bytes = pk.serialize();
-    let sha = Sha256::digest(pk_bytes);
-    let rip = Ripemd160::digest(sha);
-    let mut payload = Vec::with_capacity(25);
-    payload.push(0x00);
-    payload.extend_from_slice(&rip);
-    let check = Sha256::digest(Sha256::digest(&payload));
-    payload.extend_from_slice(&check[..4]);
-    bs58::encode(payload).into_string()
-}
-
-pub trait TransactionExt {
-    fn hash(&self) -> String;
-    fn sign(&mut self, sk: &secp256k1::SecretKey);
-    fn verify(&self) -> bool;
-}
-
-impl TransactionExt for Transaction {
-    fn hash(&self) -> String {
-        let mut hasher = Sha256::new();
-        hasher.update(self.sender.as_bytes());
-        hasher.update(self.recipient.as_bytes());
-        hasher.update(self.amount.to_be_bytes());
-        for inp in &self.inputs {
-            hasher.update(inp.address.as_bytes());
-            hasher.update(inp.amount.to_be_bytes());
-        }
-        for out in &self.outputs {
-            hasher.update(out.address.as_bytes());
-            hasher.update(out.amount.to_be_bytes());
-        }
-        hasher.update(self.fee.to_be_bytes());
-        hasher.update(&self.encrypted_message);
-        let result = hasher.finalize();
-        hex::encode(result)
-    }
-
-    fn sign(&mut self, sk: &secp256k1::SecretKey) {
-        let secp = secp256k1::Secp256k1::new();
-        let msg_hash = Sha256::digest(self.hash().as_bytes());
-        let msg = secp256k1::Message::from_slice(&msg_hash).expect("32 bytes");
-        let sig = secp.sign_ecdsa_recoverable(&msg, sk);
-        let (rec_id, data) = sig.serialize_compact();
-        self.signature.clear();
-        self.signature.push(rec_id.to_i32() as u8);
-        self.signature.extend_from_slice(&data);
-    }
-
-    fn verify(&self) -> bool {
-        if self.signature.len() != 65 {
-            return false;
-        }
-        let rec_id = match secp256k1::ecdsa::RecoveryId::from_i32(self.signature[0] as i32) {
-            Ok(id) => id,
-            Err(_) => return false,
-        };
-        let mut data = [0u8; 64];
-        data.copy_from_slice(&self.signature[1..]);
-        let sig = match secp256k1::ecdsa::RecoverableSignature::from_compact(&data, rec_id) {
-            Ok(s) => s,
-            Err(_) => return false,
-        };
-        let secp = secp256k1::Secp256k1::new();
-        let msg_hash = Sha256::digest(self.hash().as_bytes());
-        let msg = secp256k1::Message::from_slice(&msg_hash).expect("32 bytes");
-        let pk = match secp.recover_ecdsa(&msg, &sig) {
-            Ok(p) => p,
-            Err(_) => return false,
-        };
-        // derive address from recovered public key
-        let pk_bytes = pk.serialize();
-        let sha = Sha256::digest(pk_bytes);
-        let rip = Ripemd160::digest(sha);
-        let mut payload = Vec::with_capacity(25);
-        payload.push(0x00);
-        payload.extend_from_slice(&rip);
-        let check = Sha256::digest(Sha256::digest(&payload));
-        payload.extend_from_slice(&check[..4]);
-        let addr = bs58::encode(payload).into_string();
-        addr == self.sender
-    }
-}
-
-#[derive(Clone, Debug, PartialEq, serde::Serialize, serde::Deserialize)]
-pub struct BlockHeader {
-    pub previous_hash: String,
-    pub merkle_root: String,
-    /// Milliseconds since the Unix epoch
-    pub timestamp: u64,
-    pub nonce: u64,
-    pub difficulty: u32,
-}
-
-#[derive(Clone, Debug, PartialEq, serde::Serialize, serde::Deserialize)]
-pub struct Block {
-    pub header: BlockHeader,
-    pub transactions: Vec<Transaction>,
-}
-
-impl Block {
-    /// Hash the block header for block identification. The Merkle root in the
-    /// header commits to all transactions, so the transactions themselves are
-    /// not included here. This allows pruning without changing the block hash.
-    pub fn hash(&self) -> String {
-        let mut hasher = Sha256::new();
-        hasher.update(self.header.previous_hash.as_bytes());
-        hasher.update(self.header.merkle_root.as_bytes());
-        hasher.update(self.header.timestamp.to_be_bytes());
-        hasher.update(self.header.nonce.to_be_bytes());
-        hasher.update(self.header.difficulty.to_be_bytes());
-        hex::encode(hasher.finalize())
-    }
-
-    pub fn to_rpc(&self) -> coin_proto::Block {
-        coin_proto::Block {
-            header: coin_proto::BlockHeader {
-                previous_hash: self.header.previous_hash.clone(),
-                merkle_root: self.header.merkle_root.clone(),
-                timestamp: self.header.timestamp,
-                nonce: self.header.nonce,
-                difficulty: self.header.difficulty,
-            },
-            transactions: self.transactions.clone(),
-        }
-    }
-
-    pub fn from_rpc(pb: coin_proto::Block) -> Option<Self> {
-        Some(Block {
-            header: BlockHeader {
-                previous_hash: pb.header.previous_hash,
-                merkle_root: pb.header.merkle_root,
-                timestamp: pb.header.timestamp,
-                nonce: pb.header.nonce,
-                difficulty: pb.header.difficulty,
-            },
-            transactions: pb.transactions,
-        })
-    }
-
-    /// Remove transactions from the block while preserving the header. This
-    /// simulates pruning after enough confirmations.
-    pub fn prune(&mut self) {
-        self.transactions.clear();
-    }
-}
-
-pub struct Blockchain {
-    chain: Vec<Block>,
-    mempool: Vec<Transaction>,
-    difficulty: u32,
-    runtime: contract_runtime::Runtime,
-    utxos: HashMap<String, u64>,
-    locked: HashMap<String, u64>,
-}
-
-impl Blockchain {
-    pub fn new() -> Self {
-        Self {
-            chain: Vec::new(),
-            mempool: Vec::new(),
-            runtime: contract_runtime::Runtime::default(),
-            difficulty: 1,
-            utxos: HashMap::new(),
-            locked: HashMap::new(),
-        }
-    }
-
-    /// Add a transaction to the mempool if addresses are valid.
-    /// Returns `true` if the transaction was accepted.
-    pub fn add_transaction(&mut self, tx: Transaction) -> bool {
-        if (!tx.sender.is_empty() && !valid_address(&tx.sender)) || !valid_address(&tx.recipient) {
-            return false;
-        }
-        for out in &tx.outputs {
-            if !valid_address(&out.address) {
-                return false;
-            }
-        }
-        if tx.amount == 0 && tx.outputs.iter().all(|o| o.amount == 0) {
-            return false;
-        }
-        if !tx.sender.is_empty() {
-            if !tx.verify() {
-                return false;
-            }
-            if tx.inputs.is_empty() {
-                let bal = self.available_utxo(&tx.sender);
-                let total_out: u64 = tx.amount + tx.outputs.iter().map(|o| o.amount).sum::<u64>();
-                if bal < (total_out + tx.fee) as i64 {
-                    return false;
-                }
-            } else {
-                let mut sum_inputs = 0u64;
-                for inp in &tx.inputs {
-                    if !valid_address(&inp.address) {
-                        return false;
-                    }
-                    let bal = self.available_utxo(&inp.address);
-                    if bal < inp.amount as i64 {
-                        return false;
-                    }
-                    sum_inputs += inp.amount;
-                }
-                let total_out: u64 =
-                    tx.amount + tx.outputs.iter().map(|o| o.amount).sum::<u64>() + tx.fee;
-                if sum_inputs < total_out {
-                    return false;
-                }
-            }
-        }
-        self.mempool.push(tx);
-        true
-    }
-
-    /// Create a block from current mempool transactions without clearing them
-    pub fn candidate_block(&self) -> Block {
-        let previous_hash = self.last_block_hash().unwrap_or_default();
-        let merkle_root = compute_merkle_root(&self.mempool);
-        Block {
-            header: BlockHeader {
-                previous_hash,
-                merkle_root,
-                timestamp: SystemTime::now()
-                    .duration_since(UNIX_EPOCH)
-                    .unwrap_or_else(|_| std::time::Duration::new(0, 0))
-                    .as_millis() as u64,
-                nonce: 0,
-                difficulty: self.difficulty,
-            },
-            transactions: self.mempool.clone(),
-        }
-    }
-
-    /// Total amount of coins mined so far
-    pub fn total_mined(&self) -> u64 {
-        self.chain
-            .iter()
-            .flat_map(|b| &b.transactions)
-            .filter(|tx| tx.sender.is_empty())
-            .map(|tx| tx.amount)
-            .sum()
-    }
-
-    /// Calculate the block subsidy for the given height and mined amount
-    pub fn subsidy_for_height(height: u64, mined: u64) -> u64 {
-        let halvings = height / HALVING_INTERVAL;
-        if halvings >= 64 {
-            return 0;
-        }
-        let mut reward = BLOCK_SUBSIDY >> halvings;
-        if mined >= MAX_SUPPLY {
-            reward = 0;
-        } else if mined + reward > MAX_SUPPLY {
-            reward = MAX_SUPPLY - mined;
-        }
-        reward
-    }
-
-    /// Subsidy for the next block mined on this chain
-    pub fn block_subsidy(&self) -> u64 {
-        Self::subsidy_for_height(self.chain.len() as u64, self.total_mined())
-    }
-
-    /// Append a confirmed block to the chain and clear contained transactions
-    pub fn add_block(&mut self, block: Block) {
-        for tx in &block.transactions {
-            if let Some(pos) = self.mempool.iter().position(|m| m == tx) {
-                self.mempool.remove(pos);
-            }
-
-            if tx.inputs.is_empty() {
-                if !tx.sender.is_empty() {
-                    let spend =
-                        tx.amount + tx.outputs.iter().map(|o| o.amount).sum::<u64>() + tx.fee;
-                    let entry = self.utxos.entry(tx.sender.clone()).or_insert(0);
-                    *entry = entry.saturating_sub(spend);
-                }
-            } else {
-                for inp in &tx.inputs {
-                    let entry = self.utxos.entry(inp.address.clone()).or_insert(0);
-                    *entry = entry.saturating_sub(inp.amount);
-                }
-            }
-
-            let entry = self.utxos.entry(tx.recipient.clone()).or_insert(0);
-            *entry += tx.amount;
-            for out in &tx.outputs {
-                let e = self.utxos.entry(out.address.clone()).or_insert(0);
-                *e += out.amount;
-            }
-
-            if tx.recipient.is_empty() && !tx.encrypted_message.is_empty() {
-                if let Ok(dep) =
-                    serde_json::from_slice::<contract_runtime::DeployPayload>(&tx.encrypted_message)
-                {
-                    let _ = self.runtime.deploy(&tx.sender, &dep.wasm);
-                } else if let Ok(inv) =
-                    serde_json::from_slice::<contract_runtime::InvokePayload>(&tx.encrypted_message)
-                {
-                    let mut gas = 1_000_000;
-                    let _ = self.runtime.execute(&inv.contract, &mut gas);
-                }
-            }
-        }
-        self.chain.push(block);
-
-        if self.chain.len() >= DIFFICULTY_WINDOW {
-            let window = &self.chain[self.chain.len() - DIFFICULTY_WINDOW..];
-            let mut total = 0u64;
-            let mut count = 0u64;
-            for pair in window.windows(2) {
-                total += pair[1].header.timestamp - pair[0].header.timestamp;
-                count += 1;
-            }
-            if count > 0 {
-                let avg = total / count;
-                if avg < TARGET_BLOCK_MS {
-                    self.difficulty = self.difficulty.saturating_add(1);
-                } else if avg > TARGET_BLOCK_MS {
-                    self.difficulty = self.difficulty.saturating_sub(1);
-                }
-            }
-        }
-    }
-
-    pub fn last_block_hash(&self) -> Option<String> {
-        self.chain.last().map(|b| b.hash())
-    }
-
-    pub fn len(&self) -> usize {
-        self.chain.len()
-    }
-
-    pub fn difficulty(&self) -> u32 {
-        self.difficulty
-    }
-
-    pub fn validate_chain(blocks: &[Block]) -> bool {
-        let mut mined = 0u64;
-        for (i, block) in blocks.iter().enumerate() {
-            if i > 0 && block.header.previous_hash != blocks[i - 1].hash() {
-                return false;
-            }
-
-            if !block.transactions.is_empty() {
-                let mut fees = 0u64;
-                let mut coinbase: Option<&Transaction> = None;
-                for tx in &block.transactions {
-                    if tx.sender.is_empty() {
-                        if coinbase.is_some() {
-                            return false;
-                        }
-                        if tx.fee != 0 || !tx.inputs.is_empty() {
-                            return false;
-                        }
-                        coinbase = Some(tx);
-                    } else {
-                        if !tx.verify() {
-                            return false;
-                        }
-                        fees += tx.fee;
-                    }
-                }
-                let reward = Blockchain::subsidy_for_height(i as u64, mined);
-                let expected = reward + fees;
-                let cb = match coinbase {
-                    Some(c) => c,
-                    None => return false,
-                };
-                if cb.amount != expected {
-                    return false;
-                }
-                mined = mined.saturating_add(reward);
-
-                let merkle = compute_merkle_root(&block.transactions);
-                if merkle != block.header.merkle_root {
-                    return false;
-                }
-            }
-
-            if let Ok(hash) = hex::decode(block.hash()) {
-                if !meets_difficulty(&hash, block.header.difficulty) {
-                    return false;
-                }
-            } else {
-                return false;
-            }
-        }
-        true
-    }
-
-    /// Calculate the cumulative difficulty for a slice of blocks.
-    pub fn total_difficulty_of(blocks: &[Block]) -> u64 {
-        blocks.iter().map(|b| b.header.difficulty as u64).sum()
-    }
-
-    /// Total difficulty of the current chain.
-    pub fn total_difficulty(&self) -> u64 {
-        Self::total_difficulty_of(&self.chain)
-    }
-
-    pub fn replace(&mut self, new_chain: Vec<Block>) {
-        let new_diff = Self::total_difficulty_of(&new_chain);
-        let cur_diff = self.total_difficulty();
-        if new_diff > cur_diff || (new_diff == cur_diff && new_chain.len() > self.chain.len()) {
-            self.chain = new_chain;
-        }
-    }
-
-    pub fn all(&self) -> Vec<Block> {
-        self.chain.clone()
-    }
-
-    pub fn save<P: AsRef<Path>>(&self, dir: P) -> std::io::Result<()> {
-        use rocksdb::Options;
-        use std::fs;
-
-        let dir = dir.as_ref();
-        fs::create_dir_all(dir)?;
-        if blockfile::db_exists(dir) {
-            let _ = rocksdb::DB::destroy(&Options::default(), dir);
-        }
-        // remove any leftover files (including legacy files)
-        for entry in fs::read_dir(dir)? {
-            let entry = entry?;
-            if let Some(name) = entry.file_name().to_str() {
-                if name.starts_with("blk") && name.ends_with(".dat") {
-                    let _ = fs::remove_file(entry.path());
-                } else if name != "." && name != ".." {
-                    let _ = fs::remove_file(entry.path());
-                }
-            }
-        }
-        for block in &self.chain {
-            blockfile::append_block(dir, block)?;
-        }
-        utxofile::save_utxos(dir.join("utxos.bin"), &self.utxos)?;
-        Ok(())
-    }
-
-    pub fn save_mempool<P: AsRef<Path>>(&self, path: P) -> std::io::Result<()> {
-        let data = serde_json::to_vec(&self.mempool)
-            .map_err(|e| std::io::Error::new(std::io::ErrorKind::Other, e))?;
-        let mut f = File::create(path)?;
-        f.write_all(&data)
-    }
-
-    pub fn load<P: AsRef<Path>>(dir: P) -> std::io::Result<Self> {
-        let dir = dir.as_ref();
-        let blocks = if blockfile::db_exists(dir) {
-            blockfile::read_blocks(dir)?
-        } else {
-            blockfile::migrate_from_files(dir)?
-        };
-        if !Self::validate_chain(&blocks) {
-            return Err(std::io::Error::new(
-                std::io::ErrorKind::InvalidData,
-                "invalid chain",
-            ));
-        }
-        let saved_utxos = if blockfile::db_exists(dir) {
-            utxofile::load_utxos(dir.join("utxos.bin")).ok()
-        } else {
-            let map = utxofile::load_utxos(dir.join("utxos.bin")).ok();
-            if let Some(ref m) = map {
-                let _ = utxofile::save_utxos(dir.join("utxos.bin"), m);
-            }
-            map
-        };
-        let mut bc = Blockchain::new();
-        for block in &blocks {
-            bc.add_block(block.clone());
-        }
-        if let Some(map) = saved_utxos {
-            bc.utxos = map;
-        }
-        Ok(bc)
-    }
-
-    pub fn load_mempool<P: AsRef<Path>>(&mut self, path: P) -> std::io::Result<()> {
-        let f = File::open(path)?;
-        let mempool: Vec<Transaction> = serde_json::from_reader(f)
-            .map_err(|e| std::io::Error::new(std::io::ErrorKind::InvalidData, e))?;
-        self.mempool = mempool;
-        Ok(())
-    }
-
-    /// Calculate the balance for `addr` by scanning the chain
-    pub fn balance(&self, addr: &str) -> i64 {
-        let total = *self.utxos.get(addr).unwrap_or(&0) as i64;
-        let locked = *self.locked.get(addr).unwrap_or(&0) as i64;
-        total - locked
-    }
-
-    fn available_utxo(&self, addr: &str) -> i64 {
-        let mut bal = self.balance(addr);
-        for m in &self.mempool {
-            if m.inputs.is_empty() {
-                if m.sender == addr && !m.sender.is_empty() {
-                    let extra: u64 = m.outputs.iter().map(|o| o.amount).sum();
-                    bal -= (m.amount + extra + m.fee) as i64;
-                }
-            } else {
-                for inp in &m.inputs {
-                    if inp.address == addr {
-                        bal -= inp.amount as i64;
-                    }
-                }
-            }
-        }
-        bal
-    }
-
-    pub fn lock_stake(&mut self, addr: &str, amount: u64) -> bool {
-        let bal = self.balance(addr);
-        if bal < amount as i64 {
-            return false;
-        }
-        let entry = self.locked.entry(addr.to_string()).or_default();
-        *entry += amount;
-        true
-    }
-
-    pub fn unlock_stake(&mut self, addr: &str, amount: u64) -> bool {
-        if let Some(entry) = self.locked.get_mut(addr) {
-            if *entry >= amount {
-                *entry -= amount;
-                if *entry == 0 {
-                    self.locked.remove(addr);
-                }
-                return true;
-            }
-        }
-        false
-    }
-
-    pub fn mempool_len(&self) -> usize {
-        self.mempool.len()
-    }
-
-    pub fn locked_balance(&self, addr: &str) -> u64 {
-        *self.locked.get(addr).unwrap_or(&0)
-    }
-
-    /// Prune transactions from blocks older than `depth` from the tip.
-    pub fn prune(&mut self, depth: usize) {
-        let len = self.chain.len();
-        if len <= depth {
-            return;
-        }
-        for block in &mut self.chain[..len - depth] {
-            block.prune();
-        }
-    }
-}
-
-#[cfg(test)]
 mod tests {
     use super::*;
     use proptest::prelude::*;
